Add cancelOnDropOutside flag to improve UX on manual sorting
When Tree component is used to manual sort items, the user might drag
items outside of the droppable area -- especially when trying to move it
to the top or the bottom of the list. If we restrict the drop to only
the tree area, we might have some weird behaviors, like rendering the
placeholder somewhere, but not accepting a drop when the user releases
the mouse button for example.

The default is still only allowing drop inside the tree area, but by
setting `cancelOnDropOutside` as `false` we allow drops outside of the
tree whenever there's a placeholder being rendered.